CreateAndDelete

+++ routes/web.php
@@ -1,13 +1,16 @@
 
 <?php
 
+use App\Http\Middleware\IsAdmin;
+use App\Http\Middleware\IsModerador;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\HomeController;
-use App\Http\Controllers\UserController;
 use App\Http\Controllers\TemaController;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ForumController;
 use App\Http\Controllers\DesignController;
+use App\Http\Middleware\IsAdminOrModerator;
 use App\Http\Controllers\MaterialController;
 use App\Http\Controllers\DashboardController;
 
@@ -18,7 +21,7 @@ Route::get('/', function () {
 
 // Rota de fallback
 Route::fallback(function () {
-    return view('falback');
+    return view('fallback');
 });
 
 
@@ -45,17 +48,31 @@ Route::middleware('auth')->group(function () {
 
         Route::get('/', [TemaController::class, 'index'])->name('tema');
 
-        //rota para a página de adicionar temas, feita para adicionar temas, leva para o formulário de adição
-        Route::get('/add', [TemaController::class, 'addTema'])->name('tema.add');
+        Route::middleware([IsAdmin::class])->group(function(){
+            //rota para a página de adicionar temas, feita para adicionar temas, leva para o formulário de adição
+            Route::get('/add', [TemaController::class, 'addTema'])->name('tema.add');
 
-        //rota para colocar as temas que criamos no formulário no base de dados
-        Route::post('/create', [TemaController::class, 'createTema'])->name('tema.create');
+            //rota para colocar as temas que criamos no formulário no base de dados
+            Route::post('/create', [TemaController::class, 'createTema'])->name('tema.create');
+
+            //rota para deletar temas
+            Route::delete('/delete/{id}', [TemaController::class, 'deleteTema'])->name('tema.delete');
+
+
+            Route::delete('delete-multiple', [TemaController::class, 'deleteTema'])->name('tema.deleteMultiple');
+        });
+
 
         //rota para deletar temas
         Route::get('/delete/{id}', [TemaController::class, 'deleteTema'])->name('tema.delete');
+
     });
 
+    Route::get('/tema/{id}/material', [MaterialController::class,'showMaterial'])->name('tema.material');
 
+    Route::get('/material/tema/{id}', [MaterialController::class, 'showMaterial'])->name('material.show');
+
+    Route::get('/material/detalhes/{id}', [MaterialController::class, 'showDetalhes'])->name('material.detalhes');
 
     Route::prefix('material')->group(function () {
 
@@ -67,11 +84,13 @@ Route::middleware('auth')->group(function () {
         //rota para colocar os materiais que criamos no formulário na base de dados
         Route::post('/create', [MaterialController::class, 'createMaterial'])->name('material.create');
 
+
+
         // Rota para excluir um material
-        Route::delete('/delete/{id}', [MaterialController::class, 'deleteMaterial'])->name('material.delete');
+        Route::delete('/delete/{id}', [MaterialController::class, 'deleteMaterial'])->name('material.delete')->middleware([IsAdminOrModerator::class]);
 
         // Rota para excluir mais de um material ao mesmo tempo
-        Route::delete('delete-multiple', [MaterialController::class, 'deleteMaterial'])->name('material.deleteMultiple');
+        Route::delete('delete-multiple', [MaterialController::class, 'deleteMaterial'])->name('material.deleteMultiple')->middleware([IsAdminOrModerator::class]);
     });
